pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-py-repl"
version = "0.1.4"
description = "MCP server providing a Python REPL with persistent session"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.2.10",
    "pandas",
    "matplotlib",
    "seaborn",
		"asyncio",
		"traceloop-sdk",
]
readme = "README.md"

[project.scripts]
mcp-py-repl = "mcp_py_repl:main"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_py_repl"]

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["mcp_py_repl"]
